# src/web/openapi/running-performance.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RunningPerformance.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/races": {
      "get": {
        "tags": [
          "Races"
        ],
        "operationId": "GetRaces",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TargetRaceResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Races"
        ],
        "operationId": "CreateRace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTargetRaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetRaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/races/{id}": {
      "put": {
        "tags": [
          "Races"
        ],
        "operationId": "UpdateRace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTargetRaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetRaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/races/{id}/goals": {
      "get": {
        "tags": [
          "Races"
        ],
        "operationId": "GetRaceGoals",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RaceGoalResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Races"
        ],
        "operationId": "CreateRaceGoal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRaceGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceGoalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRaceGoalRequest": {
        "required": [
          "goalTimeSeconds",
          "goalPaceSecondsPerKm",
          "confidence",
          "rationale"
        ],
        "type": "object",
        "properties": {
          "goalTimeSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "goalPaceSecondsPerKm": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "confidence": {
            "type": [
              "null",
              "string"
            ]
          },
          "rationale": {
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "RaceGoalResponse": {
        "required": [
          "id",
          "versionNumber",
          "goalTimeSeconds",
          "goalPaceSecondsPerKm",
          "confidence",
          "rationale",
          "supersedesId",
          "effectiveAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "versionNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "goalTimeSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "goalPaceSecondsPerKm": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "confidence": {
            "type": [
              "null",
              "string"
            ]
          },
          "rationale": {
            "type": "string"
          },
          "supersedesId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "effectiveAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SaveTargetRaceRequest": {
        "required": [
          "name",
          "raceDate",
          "distanceM",
          "location",
          "priority",
          "status",
          "timezoneName"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "raceDate": {
            "type": "string",
            "format": "date"
          },
          "distanceM": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "priority": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "timezoneName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TargetRaceResponse": {
        "required": [
          "id",
          "name",
          "raceDate",
          "distanceM",
          "location",
          "priority",
          "status",
          "timezoneName",
          "updatedAt",
          "currentGoal"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "raceDate": {
            "type": "string",
            "format": "date"
          },
          "distanceM": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "priority": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "timezoneName": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "currentGoal": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RaceGoalResponse"
              }
            ]
          }
        }
      },
...
    }
  },
...
}